Модуляция выполнено
